+++ libraries/wallet/include/graphene/wallet/wallet.hpp
@@ -234,7 +234,7 @@ namespace graphene { namespace wallet {
          std::string         author_account;
          uint32_t            times_bought;
          fc::ripemd160       hash;
-         uint64_t            AVG_rating;
+         uint32_t            AVG_rating;
       };
 
       struct rating_object_ex : public buying_object
